backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.5.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
httpx>=0.26.0
email-validator>=2.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0